requirements.txt
python-telegram-bot[ext]
flask[async]
python-dotenv
requests
pillow